+++ app/commands/characters_context/dnd5/spell_add_command.rb
@@ -6,10 +6,13 @@ module CharactersContext
       use_contract do
         config.messages.namespace = :dnd5_character
 
+        SpellAbilities = Dry::Types['strict.string'].enum('int', 'wis', 'cha')
+
         params do
           required(:character).filled(type?: ::Dnd5::Character)
           required(:spell).filled(type?: ::Dnd5::Spell)
           required(:target_spell_class).filled(:string)
+          optional(:spell_ability).filled(SpellAbilities)
         end
       end
 
@@ -21,8 +24,9 @@ module CharactersContext
           spell: input[:spell],
           data: {
             prepared_by: input[:target_spell_class],
-            ready_to_use: input[:target_spell_class] != ::Dnd5::Character::WIZARD
-          }
+            ready_to_use: input[:target_spell_class] != ::Dnd5::Character::WIZARD,
+            spell_ability: input[:spell_ability]
+          }.compact
         )
 
         { result: result }