+++ app/controllers/frontend/dnd2024/characters/spells_controller.rb
@@ -12,16 +12,17 @@ module Frontend
         include SerializeRelation
         include SerializeResource
 
-        INDEX_SERIALIZER_FIELDS = %i[id ready_to_use prepared_by slug name level spell_id].freeze
+        INDEX_SERIALIZER_FIELDS = %i[id ready_to_use prepared_by spell_ability notes slug name level spell_id].freeze
 
         before_action :find_character
+        before_action :find_spell, only: %i[create]
 
         def index
           serialize_relation(spells, ::Dnd2024::Characters::SpellSerializer, :spells, only: INDEX_SERIALIZER_FIELDS)
         end
 
         def create
-          case character_spell_add.call(create_params)
+          case character_spell_add.call(create_params.merge(character: @character, spell: @spell))
           in { errors: errors, errors_list: errors_list } then unprocessable_response(errors, errors_list)
           in { result: result }
             serialize_resource(result, ::Dnd2024::Characters::SpellSerializer, :spell, {}, :created)
@@ -29,7 +30,7 @@ module Frontend
         end
 
         def update
-          case character_spell_update.call(update_params)
+          case character_spell_update.call(update_params.merge(character: @character))
           in { errors: errors, errors_list: errors_list } then unprocessable_response(errors, errors_list)
           else only_head_response
           end
@@ -47,21 +48,20 @@ module Frontend
           @character = authorized_scope(Character.all).dnd2024.find(params[:character_id])
         end
 
+        def find_spell
+          @spell = ::Spell.dnd2024.find(params[:spell_id])
+        end
+
         def spells
           @character.spells.includes(:spell)
         end
 
         def create_params
-          {
-            character: @character,
-            spell: ::Spell.dnd2024.find(params[:spell_id]),
-            target_spell_class: params[:target_spell_class]
-          }
+          params.permit(:target_spell_class, :spell_ability).to_h
         end
 
         def update_params
           {
-            character: @character,
             character_spell: @character.spells.find(params[:id]),
             ready_to_use: to_bool.call(params[:ready_to_use])
           }