+++ app/decorators/dnd2024_character/classes/rogue_decorator.rb
@@ -15,6 +15,18 @@ module Dnd2024Character
           end
       end
 
+      def spell_classes
+        @spell_classes ||= begin
+          result = __getobj__.spell_classes
+          result[:rogue] = { multiclass_spell_level: 0 }
+          result
+        end
+      end
+
+      def spells_slots
+        @spells_slots ||= ::Dnd2024Character::ClassDecorateWrapper::EMPTY_SPELL_SLOTS[class_level]
+      end
+
       private
 
       def class_level