+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - character personal homebrew feats
 - add weapons for feats to generating new attacks
 - selecting backgrounds for new D&D 2024 characters
+- selecting spell casting ability while learning spells for D&D characters
 
 ### Modified
 - rendering basic PDF for all systems